pyproject.toml
[project]
name = "paper-readings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.5",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = []
